Resolves #903 Creating a bunch more fixtures
This probably isn't perfect, but better, I hope.
There were no code changes to any of the methods as part of this
change, just moving them between classes.
