+++ fdb-record-layer-core/src/test/java/com/apple/foundationdb/record/provider/foundationdb/FDBRecordStoreUniquenessTest.java
@@ -1,0 +1,114 @@
+/*
+ * FDBRecordStoreUniquenessTest.java
+ *
+ * This source file is part of the FoundationDB open source project
+ *
+ * Copyright 2015-2020 Apple Inc. and the FoundationDB project authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.apple.foundationdb.record.provider.foundationdb;
+
+import com.apple.foundationdb.record.RecordIndexUniquenessViolation;
+import com.apple.foundationdb.record.TestRecords1Proto;
+import com.apple.foundationdb.record.TestRecordsBytesProto;
+import com.apple.foundationdb.tuple.Tuple;
+import com.apple.test.Tags;
+import org.junit.jupiter.api.Tag;
+import org.junit.jupiter.api.Test;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Random;
+import java.util.concurrent.CompletableFuture;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+
+/**
+ * Tests of uniqueness checks.
+ */
+@Tag(Tags.RequiresFDB)
+public class FDBRecordStoreUniquenessTest extends FDBRecordStoreTestBase {
+
+    @Test
+    public void writeUniqueByteString() throws Exception {
+        try (FDBRecordContext context = openContext()) {
+            openBytesRecordStore(context);
+
+            recordStore.saveRecord(TestRecordsBytesProto.ByteStringRecord.newBuilder()
+                    .setPkey(byteString(0, 1, 2)).setSecondary(byteString(0, 1, 2)).setUnique(byteString(0, 2))
+                    .setName("foo").build());
+            recordStore.saveRecord(TestRecordsBytesProto.ByteStringRecord.newBuilder()
+                    .setPkey(byteString(0, 1, 5)).setSecondary(byteString(0, 1, 3)).setUnique(byteString(0, 2))
+                    .setName("box").build());
+            assertThrows(RecordIndexUniquenessViolation.class, () -> commit(context));
+        }
+    }
+
+    @Test
+    public void asyncUniqueInserts() throws Exception {
+        List<TestRecords1Proto.MySimpleRecord> records = new ArrayList<>();
+        Random r = new Random(0xdeadc0deL);
+        for (int i = 0; i < 100; i++) {
+            records.add(TestRecords1Proto.MySimpleRecord.newBuilder()
+                    .setRecNo(r.nextLong())
+                    .setNumValueUnique(r.nextInt())
+                    .build()
+            );
+        }
+
+        try (FDBRecordContext context = openContext()) {
+            openSimpleRecordStore(context);
+
+            CompletableFuture<?>[] futures = new CompletableFuture<?>[records.size()];
+            for (int i = 0; i < records.size(); i++) {
+                futures[i] = recordStore.saveRecordAsync(records.get(i));
+            }
+
+            CompletableFuture.allOf(futures).get();
+            commit(context);
+        }
+
+        try (FDBRecordContext context = openContext()) {
+            openSimpleRecordStore(context);
+
+            for (TestRecords1Proto.MySimpleRecord record : records) {
+                assertEquals(record.toString(), recordStore.loadRecord(Tuple.from(record.getRecNo())).getRecord().toString());
+            }
+        }
+    }
+
+    @Test
+    public void asyncNotUniqueInserts() throws Exception {
+        try (FDBRecordContext context = openContext()) {
+            openSimpleRecordStore(context);
+
+            CompletableFuture<?>[] futures = new CompletableFuture<?>[2];
+            futures[0] = recordStore.saveRecordAsync(TestRecords1Proto.MySimpleRecord.newBuilder()
+                    .setRecNo(1066L)
+                    .setNumValueUnique(42)
+                    .build()
+            );
+            futures[1] = recordStore.saveRecordAsync(TestRecords1Proto.MySimpleRecord.newBuilder()
+                    .setRecNo(1776L)
+                    .setNumValueUnique(42)
+                    .build()
+            );
+
+            CompletableFuture.allOf(futures).get();
+            assertThrows(RecordIndexUniquenessViolation.class, () -> commit(context));
+        }
+    }
+}
